# Improve `create_input` and `create_output`