--- a/crates/brioche-core/benches/bake.rs
+++ b/crates/brioche-core/benches/bake.rs
@@ -2,8 +2,172 @@ use brioche_core::{
     recipe::{Directory, Recipe, WithMeta},
     Brioche,
 };
-use criterion::{criterion_group, criterion_main, Criterion};
-use futures::StreamExt as _;
+
+fn main() {
+    divan::main();
+}
+
+#[divan::bench]
+fn cached_bake_deep_dir(bencher: divan::Bencher) {
+    let RecipesContext {
+        brioche,
+        runtime,
+        recipes,
+        ..
+    } = set_up_bench();
+
+    bencher.bench_local(|| {
+        runtime.block_on(async {
+            let deep_dir = WithMeta::without_meta(Recipe::from(recipes.deep_dir.clone()));
+            let _ = brioche_core::bake::bake(
+                &brioche,
+                deep_dir,
+                &brioche_core::bake::BakeScope::Anonymous,
+            )
+            .await
+            .unwrap();
+        });
+    })
+}
+
+#[divan::bench]
+fn cached_bake_wide_dir(bencher: divan::Bencher) {
+    let RecipesContext {
+        brioche,
+        runtime,
+        recipes,
+        ..
+    } = set_up_bench();
+
+    bencher.bench_local(|| {
+        runtime.block_on(async {
+            let wide_dir = WithMeta::without_meta(Recipe::from(recipes.wide_dir.clone()));
+            let _ = brioche_core::bake::bake(
+                &brioche,
+                wide_dir,
+                &brioche_core::bake::BakeScope::Anonymous,
+            )
+            .await
+            .unwrap();
+        });
+    })
+}
+
+#[divan::bench]
+fn cached_bake_deep_merge(bencher: divan::Bencher) {
+    let RecipesContext {
+        brioche,
+        runtime,
+        recipes,
+        ..
+    } = set_up_bench();
+
+    bencher.bench_local(|| {
+        runtime.block_on(async {
+            let merge_deep_dir = WithMeta::without_meta(recipes.merge_deep_dir.clone());
+            let _ = brioche_core::bake::bake(
+                &brioche,
+                merge_deep_dir,
+                &brioche_core::bake::BakeScope::Anonymous,
+            )
+            .await
+            .unwrap();
+        });
+    })
+}
+
+#[divan::bench]
+fn cached_bake_wide_merge(bencher: divan::Bencher) {
+    let RecipesContext {
+        brioche,
+        runtime,
+        recipes,
+        ..
+    } = set_up_bench();
+
+    bencher.bench_local(|| {
+        runtime.block_on(async {
+            let merge_wide_dir = WithMeta::without_meta(recipes.merge_wide_dir.clone());
+            let _ = brioche_core::bake::bake(
+                &brioche,
+                merge_wide_dir,
+                &brioche_core::bake::BakeScope::Anonymous,
+            )
+            .await
+            .unwrap();
+        });
+    })
+}
+
+fn set_up_bench() -> RecipesContext {
+    let runtime = tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()
+        .unwrap();
+    let (brioche, context) = runtime.block_on(brioche_test_support::brioche_test());
+
+    let deep_dir = runtime.block_on(make_deep_dir(&brioche, ""));
+    let _deep_dir_result = runtime
+        .block_on(brioche_core::bake::bake(
+            &brioche,
+            WithMeta::without_meta(Recipe::from(deep_dir.clone())),
+            &brioche_core::bake::BakeScope::Anonymous,
+        ))
+        .unwrap();
+
+    let wide_dir = runtime.block_on(make_wide_dir(&brioche, ""));
+    let _wide_dir_result = runtime
+        .block_on(brioche_core::bake::bake(
+            &brioche,
+            WithMeta::without_meta(Recipe::from(wide_dir.clone())),
+            &brioche_core::bake::BakeScope::Anonymous,
+        ))
+        .unwrap();
+
+    let merge_deep_dir = Recipe::Merge {
+        directories: (0..10)
+            .map(|n| {
+                let deep_dir = runtime.block_on(make_deep_dir(&brioche, &n.to_string()));
+                WithMeta::without_meta(Recipe::from(deep_dir))
+            })
+            .collect(),
+    };
+
+    let merge_wide_dir = Recipe::Merge {
+        directories: (0..10)
+            .map(|n| {
+                let wide_dir = runtime.block_on(make_wide_dir(&brioche, &n.to_string()));
+                WithMeta::without_meta(Recipe::from(wide_dir))
+            })
+            .collect(),
+    };
+
+    RecipesContext {
+        brioche,
+        _context: context,
+        runtime,
+        recipes: Recipes {
+            deep_dir,
+            wide_dir,
+            merge_deep_dir,
+            merge_wide_dir,
+        },
+    }
+}
+
+struct RecipesContext {
+    brioche: Brioche,
+    _context: brioche_test_support::TestContext,
+    runtime: tokio::runtime::Runtime,
+    recipes: Recipes,
+}
+
+struct Recipes {
+    deep_dir: Directory,
+    wide_dir: Directory,
+    merge_deep_dir: Recipe,
+    merge_wide_dir: Recipe,
+}
 
 async fn make_deep_dir(brioche: &Brioche, key: &str) -> Directory {
     let mut directory = Directory::default();
@@ -59,134 +223,3 @@ async fn make_wide_dir(brioche: &Brioche, key: &str) -> Directory {
 
     directory
 }
-
-fn run_bake_benchmark(c: &mut Criterion) {
-    let runtime = tokio::runtime::Builder::new_multi_thread()
-        .enable_all()
-        .build()
-        .expect("failed to build Tokio runtime");
-    let _runtime_guard = runtime.enter();
-
-    struct Recipes {
-        deep_dir: Directory,
-        wide_dir: Directory,
-        merge_deep_dir: Recipe,
-        merge_wide_dir: Recipe,
-    }
-
-    let (brioche, _context, recipes) = runtime.block_on(async {
-        let (brioche, context) = brioche_test_support::brioche_test().await;
-
-        let deep_dir = make_deep_dir(&brioche, "").await;
-        let _deep_dir_result = brioche_core::bake::bake(
-            &brioche,
-            WithMeta::without_meta(Recipe::from(deep_dir.clone())),
-            &brioche_core::bake::BakeScope::Anonymous,
-        )
-        .await
-        .unwrap();
-
-        let wide_dir = make_wide_dir(&brioche, "").await;
-        let _wide_dir_result = brioche_core::bake::bake(
-            &brioche,
-            WithMeta::without_meta(Recipe::from(wide_dir.clone())),
-            &brioche_core::bake::BakeScope::Anonymous,
-        )
-        .await
-        .unwrap();
-
-        let merge_deep_dir = Recipe::Merge {
-            directories: futures::stream::iter(0..10)
-                .then(|n| {
-                    let brioche = brioche.clone();
-                    async move {
-                        WithMeta::without_meta(Recipe::from(
-                            make_deep_dir(&brioche, &n.to_string()).await,
-                        ))
-                    }
-                })
-                .collect()
-                .await,
-        };
-
-        let merge_wide_dir = Recipe::Merge {
-            directories: futures::stream::iter(0..10)
-                .then(|n| {
-                    let brioche = brioche.clone();
-                    async move {
-                        WithMeta::without_meta(Recipe::from(
-                            make_deep_dir(&brioche, &n.to_string()).await,
-                        ))
-                    }
-                })
-                .collect()
-                .await,
-        };
-
-        (
-            brioche,
-            context,
-            Recipes {
-                deep_dir,
-                wide_dir,
-                merge_deep_dir,
-                merge_wide_dir,
-            },
-        )
-    });
-
-    c.bench_function("cached bake deep dir", |b| {
-        b.to_async(&runtime).iter(|| async {
-            let deep_dir = WithMeta::without_meta(Recipe::from(recipes.deep_dir.clone()));
-            let _ = brioche_core::bake::bake(
-                &brioche,
-                deep_dir,
-                &brioche_core::bake::BakeScope::Anonymous,
-            )
-            .await
-            .unwrap();
-        })
-    });
-
-    c.bench_function("cached bake wide dir", |b| {
-        b.to_async(&runtime).iter(|| async {
-            let wide_dir = WithMeta::without_meta(Recipe::from(recipes.wide_dir.clone()));
-            let _ = brioche_core::bake::bake(
-                &brioche,
-                wide_dir,
-                &brioche_core::bake::BakeScope::Anonymous,
-            )
-            .await
-            .unwrap();
-        })
-    });
-
-    c.bench_function("cached bake deep merge", |b| {
-        b.to_async(&runtime).iter(|| async {
-            let merge_deep_dir = WithMeta::without_meta(recipes.merge_deep_dir.clone());
-            let _ = brioche_core::bake::bake(
-                &brioche,
-                merge_deep_dir,
-                &brioche_core::bake::BakeScope::Anonymous,
-            )
-            .await
-            .unwrap();
-        })
-    });
-
-    c.bench_function("cached bake wide merge", |b| {
-        b.to_async(&runtime).iter(|| async {
-            let merge_wide_dir = WithMeta::without_meta(recipes.merge_wide_dir.clone());
-            let _ = brioche_core::bake::bake(
-                &brioche,
-                merge_wide_dir,
-                &brioche_core::bake::BakeScope::Anonymous,
-            )
-            .await
-            .unwrap();
-        })
-    });
-}
-
-criterion_group!(benches, run_bake_benchmark);
-criterion_main!(benches);