--- a/crates/brioche-core/benches/input.rs
+++ b/crates/brioche-core/benches/input.rs
@@ -1,0 +1,167 @@
+use brioche_core::Brioche;
+
+fn main() {
+    divan::main();
+}
+
+#[divan::bench(args = [Removal::Remove, Removal::Keep])]
+fn bench_input(bencher: divan::Bencher, removal: Removal) {
+    bencher
+        .with_inputs(|| {
+            let runtime = tokio::runtime::Builder::new_current_thread()
+                .enable_all()
+                .build()
+                .unwrap();
+            let (brioche, context) = runtime.block_on(brioche_test_support::brioche_test());
+
+            let input_root = context.path(ulid::Ulid::new().to_string());
+
+            let input_path = input_root.join("input");
+            let input_resources = input_root.join("resources");
+
+            std::fs::create_dir_all(&input_path).unwrap();
+            std::fs::create_dir_all(&input_resources).unwrap();
+
+            for i in 0..100 {
+                let mut file =
+                    std::fs::File::create(input_path.join(format!("file{i}.txt"))).unwrap();
+
+                brioche_pack::inject_pack(
+                    &mut file,
+                    &brioche_pack::Pack::Metadata {
+                        resource_paths: vec![format!("{i}_resources").into()],
+                        format: "".to_string(),
+                        metadata: vec![],
+                    },
+                )
+                .unwrap();
+
+                let resource_dir = input_resources.join(format!("{i}_resources"));
+                std::fs::create_dir_all(&resource_dir).unwrap();
+
+                for j in 0..10 {
+                    std::fs::write(
+                        resource_dir.join(format!("resource{j}.txt")),
+                        format!("{i}/{j}"),
+                    )
+                    .unwrap();
+                }
+            }
+
+            InputContext {
+                runtime,
+                brioche,
+                _context: context,
+                input_path,
+                input_resources,
+            }
+        })
+        .bench_values(|ctx| {
+            ctx.runtime.block_on(async {
+                let meta = Default::default();
+                brioche_core::input::create_input(
+                    &ctx.brioche,
+                    brioche_core::input::InputOptions {
+                        input_path: &ctx.input_path,
+                        remove_input: removal.should_remove(),
+                        resource_dir: Some(&ctx.input_resources),
+                        input_resource_dirs: &[],
+                        meta: &meta,
+                    },
+                )
+                .await
+                .unwrap();
+            });
+        });
+}
+
+#[divan::bench(args = [Removal::Remove, Removal::Keep])]
+fn bench_input_with_shared_resources(bencher: divan::Bencher, removal: Removal) {
+    bencher
+        .with_inputs(|| {
+            let runtime = tokio::runtime::Builder::new_current_thread()
+                .enable_all()
+                .build()
+                .unwrap();
+            let (brioche, context) = runtime.block_on(brioche_test_support::brioche_test());
+
+            let input_root = context.path(ulid::Ulid::new().to_string());
+
+            let input_path = input_root.join("input");
+            let input_resources = input_root.join("resources");
+
+            std::fs::create_dir_all(&input_path).unwrap();
+            std::fs::create_dir_all(&input_resources).unwrap();
+
+            for i in 0..100 {
+                let mut file =
+                    std::fs::File::create(input_path.join(format!("file{i}.txt"))).unwrap();
+
+                brioche_pack::inject_pack(
+                    &mut file,
+                    &brioche_pack::Pack::Metadata {
+                        resource_paths: vec!["common_resources".into()],
+                        format: "".to_string(),
+                        metadata: vec![],
+                    },
+                )
+                .unwrap();
+            }
+
+            let resource_dir = input_resources.join("common_resources");
+            std::fs::create_dir_all(&resource_dir).unwrap();
+
+            for i in 0..10 {
+                std::fs::write(resource_dir.join(format!("resource{i}.txt")), i.to_string())
+                    .unwrap();
+            }
+
+            InputContext {
+                runtime,
+                brioche,
+                _context: context,
+                input_path,
+                input_resources,
+            }
+        })
+        .bench_values(|ctx| {
+            ctx.runtime.block_on(async {
+                let meta = Default::default();
+                brioche_core::input::create_input(
+                    &ctx.brioche,
+                    brioche_core::input::InputOptions {
+                        input_path: &ctx.input_path,
+                        remove_input: removal.should_remove(),
+                        resource_dir: Some(&ctx.input_resources),
+                        input_resource_dirs: &[],
+                        meta: &meta,
+                    },
+                )
+                .await
+                .unwrap();
+            });
+        });
+}
+
+struct InputContext {
+    runtime: tokio::runtime::Runtime,
+    brioche: Brioche,
+    _context: brioche_test_support::TestContext,
+    input_path: std::path::PathBuf,
+    input_resources: std::path::PathBuf,
+}
+
+#[derive(Debug, Clone, Copy)]
+enum Removal {
+    Remove,
+    Keep,
+}
+
+impl Removal {
+    fn should_remove(self) -> bool {
+        match self {
+            Removal::Remove => true,
+            Removal::Keep => false,
+        }
+    }
+}