… Cargo.lock
… Cargo.toml
… crates/brioche-core/Cargo.toml
… crates/brioche-core/benches/brioche_bench/mod.rs
… crates/brioche-core/src/bake.rs
… crates/brioche-core/src/bake/process.rs
… crates/brioche-core/src/bake/unarchive.rs
… crates/brioche-core/src/blob.rs
… crates/brioche-core/src/download.rs
… crates/brioche-core/src/input.rs
… crates/brioche-core/src/output.rs
… crates/brioche-core/src/project.rs
… crates/brioche-core/src/reporter.rs
… crates/brioche-core/tests/bake_basic.rs
… crates/brioche-core/tests/bake_cache.rs
… crates/brioche-core/tests/bake_collect_references.rs
… crates/brioche-core/tests/bake_download.rs
… crates/brioche-core/tests/bake_parallel.rs
… crates/brioche-core/tests/bake_process.rs
… crates/brioche-core/tests/bake_proxy.rs
… crates/brioche-core/tests/directory.rs
… crates/brioche-core/tests/input.rs
… crates/brioche-core/tests/output.rs
… crates/brioche-core/tests/project_hash.rs
… crates/brioche-core/tests/project_load.rs
… crates/brioche-core/tests/recipe.rs
… crates/brioche-core/tests/recipe_hash_stable.rs
… crates/brioche-core/tests/registry_client.rs
… crates/brioche-core/tests/script_check.rs
… crates/brioche-core/tests/script_eval.rs